// frontend/src/types/api.ts
export interface PlayerSeasonShootingStats {
  season: number;
  teamId: number;
  team: string;
  conference: string | null;
  trackedShots: number;
  assistedPct: number;
  dunks: SeasonShootingStatsDunks;
  layups: SeasonShootingStatsDunks;
  tipIns: ShotTypeBreakdown;
  twoPointJumpers: SeasonShootingStatsDunks;
  threePointJumpers: SeasonShootingStatsDunks;
  attemptsBreakdown: SeasonShootingStatsAttemptsBreakdown;
  athleteId: number;
  athleteName: string;
}
// ...
export interface SeasonShootingStats {
  season: number;
  teamId: number;
  team: string;
  conference: string | null;
  trackedShots: number;
  assistedPct: number;
  dunks: SeasonShootingStatsDunks;
  layups: SeasonShootingStatsDunks;
  tipIns: ShotTypeBreakdown;
  twoPointJumpers: SeasonShootingStatsDunks;
  threePointJumpers: SeasonShootingStatsDunks;
  attemptsBreakdown: SeasonShootingStatsAttemptsBreakdown;
}

export interface SeasonShootingStatsAttemptsBreakdown {
  threePointJumpers: number;
  twoPointJumpers: number;
  tipIns: number;
  layups: number;
  dunks: number;
}

export interface SeasonShootingStatsDunks {
  made: number;
  attempted: number;
  pct: number;
  assistedPct: number;
  assisted: number;
}
// ...
export interface ShotTypeBreakdown {
  made: number;
  attempted: number;
  pct: number;
}